refactor(github): SPEC 構造化スキーマを gwt-github の公開モジュールへ移設する
SPEC-3060 (arch-review H1)。--edit spec --json の構造化スキーマ
（StructuredSpecInput / render / merge / split / 正規化ヘルパ）は gwtd CLI
の非公開モジュールに実装されており、GUI 等の他クライアントが再利用
できなかった。純粋ロジックを gwt_github::spec_structured として公開し、
CLI には入出力（stdin / ファイル読み込み）と再エクスポートだけを残す:

- crates/gwt-github/src/spec_structured.rs を新設（全公開 API に rustdoc、
  挙動は移設前とバイト一致）
- crates/gwt-github/tests/spec_structured_test.rs を追加（render の優先度
  正規化・FR/NFR/SC 再採番、merge の canonical 置換・未知セクション保持・
  空化による削除、split、不正 JSON の SpecOpsError — 6 件、RED→GREEN）
- cli/issue_spec/structured.rs は薄いアダプタに縮小（mod.rs / tests.rs は
  無修正のまま全テスト通過 = 互換の回帰ゲート）
- SPEC_CREATE_HELP の所有者記述を gwt-github に更新
